requirements.txt
# LLM API
openai>=1.30.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web UI
flask>=3.0.0

# Progress bar
tqdm>=4.66.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=8.0.0

# Utilities (hashlib, uuid, sqlite3 are built-in)
